refactor(front): split repositories' institutions manager and dialog
